// src/api/generated/schemas/chatSession.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * The Bhakti Vault API
 * FastAPI application with OpenID Connect authentication
 * OpenAPI spec version: 1.0.0
 */
import type { ChatSessionChatTopic } from './chatSessionChatTopic';

export interface ChatSession {
  public_id: string;
  user_id: number;
  chat_topic: ChatSessionChatTopic;
  created_on: number;
  updated_on: number;
}