// src/api/generated/schemas/browsingHistory.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * The Bhakti Vault API
 * FastAPI application with OpenID Connect authentication
 * OpenAPI spec version: 1.0.0
 */

export interface BrowsingHistory {
  public_id: string;
  user_id: number;
  transcript_id: string;
  transcript_semantic_title: string;
  last_accessed_on: number;
}